refactor extract useRequestPermissionsAction
